Add release command
