+++ packages/ploys-cli/src/project/info.rs
@@ -1,12 +1,11 @@
-use std::str::FromStr;
-
-use anyhow::{anyhow, Error};
+use anyhow::Error;
 use clap::Args;
 use console::style;
-use ploys::project::source::github::{GitHub, Repository};
+use ploys::project::source::github::GitHub;
 use ploys::project::source::Source;
 use ploys::project::Project;
-use url::Url;
+
+use crate::util::repo_or_url::RepoOrUrl;
 
 /// Gets the project information.
 #[derive(Args)]
@@ -70,54 +69,5 @@ impl Info {
         }
 
         Ok(())
-    }
-}
-
-#[derive(Clone)]
-enum RepoOrUrl {
-    Repo(Repository),
-    Url(Url),
-}
-
-impl RepoOrUrl {
-    fn try_into_repo(self) -> Result<Repository, Error> {
-        self.try_into()
-    }
-}
-
-impl FromStr for RepoOrUrl {
-    type Err = Error;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s.parse::<Repository>() {
-            Ok(repo) => Ok(Self::Repo(repo)),
-            Err(_) => match s.parse::<Url>() {
-                Ok(url) => Ok(Self::Url(url)),
-                Err(_) => Err(anyhow!("Expected owner/repo or URL, found: {}", s)),
-            },
-        }
-    }
-}
-
-impl TryFrom<RepoOrUrl> for Repository {
-    type Error = Error;
-
-    fn try_from(value: RepoOrUrl) -> Result<Self, Self::Error> {
-        match value {
-            RepoOrUrl::Repo(repo) => Ok(repo),
-            RepoOrUrl::Url(url) => {
-                if url.domain() != Some("github.com") {
-                    return Err(anyhow!(
-                        "Unsupported remote repository: Only GitHub is supported"
-                    ));
-                }
-
-                Ok(url
-                    .path()
-                    .trim_start_matches('/')
-                    .trim_end_matches(".git")
-                    .parse::<Repository>()?)
-            }
-        }
     }
 }
